Only load tool once
